# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Photo Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /trash:
    get:
      operationId: listTrash
      tags: [trash]
      summary: List trashed photos
      responses:
        "200":
          description: Trashed photos
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Photo" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    delete:
      operationId: emptyTrash
      tags: [trash]
      summary: Permanently delete all trashed photos
      responses:
        "204":
          description: Trash emptied
  /trash/{id}/restore:
    post:
      operationId: restorePhoto
      tags: [trash]
      summary: Restore a trashed photo
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Restored photo
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Photo" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
  /trash/{id}:
    delete:
      operationId: permanentlyDeletePhoto
      tags: [trash]
      summary: Permanently delete a single trashed photo
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Permanently deleted
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }

  # ─── Search ────────────────────────────────────────────────────────────────
...
  /users/me:
    get:
      operationId: getCurrentUser
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserProfile" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    patch:
      operationId: updateCurrentUser
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UserProfileUpdate" }
      responses:
        "200":
          description: Updated user profile
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserProfile" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
  /users/me/storage:
    get:
      operationId: getUserStorageStats
      tags: [users]
      summary: Get storage usage statistics
      responses:
        "200":
          description: Storage usage
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StorageStats" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }

  # ─── Storage ───────────────────────────────────────────────────────────────
  /storage/uploads/request-url:
    post:
      tags: [storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for file upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadUrlRequest"
      responses:
        "200":
          description: Presigned upload URL generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResponse"
        "400":
          description: Missing or invalid required fields
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
        "500":
          description: Failed to generate upload URL
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
  /storage/objects/{objectPath}:
    get:
      tags: [storage]
      operationId: getStorageObject
      summary: Serve an uploaded object
      parameters:
        - name: objectPath
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Object content
          content:
            "*/*":
              schema: { type: string, format: binary }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
components:
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ─── Photos ──────────────────────────────────────────────────────────────
    Photo:
      type: object
      required:
        - id
        - userId
        - filename
        - objectPath
        - mimeType
        - fileSize
        - width
        - height
        - isFavorited
        - isArchived
        - isTrashed
        - tags
        - createdAt
        - updatedAt
      properties:
        id: { type: integer }
        userId: { type: string }
        filename: { type: string }
        objectPath: { type: string }
        thumbnailPath: { type: ["string", "null"] }
        mimeType: { type: string }
        fileSize: { type: integer }
        width: { type: integer }
        height: { type: integer }
        isFavorited: { type: boolean }
        isArchived: { type: boolean }
        isTrashed: { type: boolean }
        trashedAt: { type: ["string", "null"], format: date-time }
        caption: { type: ["string", "null"] }
        aiTags: { type: ["string", "null"] }
        exifData: { type: ["string", "null"] }
        tags: { type: array, items: { type: string } }
        albums: { type: array, items: { $ref: "#/components/schemas/AlbumRef" } }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    AlbumRef:
      type: object
      required: [id, name]
      properties:
        id: { type: integer }
        name: { type: string }
...
    UserProfile:
      type: object
      required: [id, clerkUserId, email, createdAt]
      properties:
        id: { type: integer }
        clerkUserId: { type: string }
        email: { type: string }
        displayName: { type: ["string", "null"] }
        avatarObjectPath: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
    UserProfileUpdate:
      type: object
      properties:
        displayName: { type: string }
        avatarObjectPath: { type: string }
    StorageStats:
      type: object
      required: [totalBytes, photoCount, limitBytes]
      properties:
        totalBytes: { type: integer }
        photoCount: { type: integer }
        limitBytes: { type: integer }

    # ─── Storage ─────────────────────────────────────────────────────────────
    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name: { type: string, minLength: 1 }
        size: { type: integer, minimum: 1 }
        contentType: { type: string, minLength: 1 }
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL: { type: string }
        objectPath: { type: string }
        metadata: { $ref: "#/components/schemas/UploadUrlRequest" }