# lib/api-zod/src/generated/types/userProfileUpdate.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * AI Photo Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface UserProfileUpdate {
  displayName?: string;
  avatarObjectPath?: string;
}